requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
discord.py==2.6.4
Flask==3.1.2
fonttools==4.58.4
frozenlist==1.8.0
graphviz==0.21
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
mysql-connector-python==9.4.0
packaging==25.0
pillow==11.2.1
propcache==0.4.1
python-dotenv==1.2.1
six==1.17.0
typing_extensions==4.15.0
Werkzeug==3.1.5
yarl==1.22.0
